display baselines

+++ src/DreamPotato.Core/BitHelpers.cs
@@ -4,14 +4,14 @@ namespace DreamPotato.Core;
 
 public static class BitHelpers
 {
-    public static bool ReadBit(byte operand, byte bit)
+    public static bool ReadBit(byte operand, int bit)
     {
         Debug.Assert(bit is >= 0 and < 8);
         return (operand & (1 << bit)) != 0;
     }
 
     // TODO: simple xor simpler?
-    public static void WriteBit(ref byte dest, byte bit, bool value)
+    public static void WriteBit(ref byte dest, int bit, bool value)
     {
         Debug.Assert(bit is >= 0 and < 8);
         if (value)
@@ -20,7 +20,7 @@ public static class BitHelpers
             dest = (byte)(dest & ~(1 << bit));
     }
 
-    public static byte WithBit(byte operand, byte bit, bool value)
+    public static byte WithBit(byte operand, int bit, bool value)
     {
         WriteBit(ref operand, bit, value);
         return operand;
